requirements.txt
streamlit>=1.28.0
openai>=1.3.0
python-docx>=1.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pydantic>=2.5.0
